+++ yash-syntax/CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- `parser::lex::PartialHereDoc`
 - `source::Location::advance`
 
 ## [0.3.0] - 2022-02-06